Update greeter async server to use the new coroutine executor
