+++ python/Ice/greeter/server_asyncio/chatbot.py
@@ -1,0 +1,19 @@
+# Copyright (c) ZeroC, Inc.
+
+import asyncio
+import Ice
+
+# Slice module VisitorCenter in Greeter.ice maps to Python module VisitorCenter.
+import VisitorCenter
+
+class Chatbot(VisitorCenter.Greeter):
+    """
+    Chatbot is an Ice servant that implements Slice interface Greeter.
+    """
+
+    # Implements the method greet from the Greeter class generated by the Slice compiler.
+    # This variant is the asynchronous implementation.
+    async def greet(self, name: str, current: Ice.Current) -> str:
+        await asyncio.sleep(1)
+        print(f"Dispatching greet request {{ name = '{name}' }}")
+        return f"Hello, {name}!"
